Move classes. Javadoc.
