+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/Main.java
@@ -11,8 +11,8 @@ import io.netty.handler.codec.http.HttpResponseStatus;
 import io.strimzi.controller.cluster.operator.assembly.KafkaAssemblyOperator;
 import io.strimzi.controller.cluster.operator.assembly.KafkaConnectAssemblyOperator;
 import io.strimzi.controller.cluster.operator.assembly.KafkaConnectS2IAssemblyOperator;
-import io.strimzi.controller.cluster.operator.assembly.KafkaSetOperator;
-import io.strimzi.controller.cluster.operator.assembly.ZookeeperSetOperator;
+import io.strimzi.controller.cluster.operator.resource.KafkaSetOperator;
+import io.strimzi.controller.cluster.operator.resource.ZookeeperSetOperator;
 import io.strimzi.controller.cluster.operator.resource.BuildConfigOperator;
 import io.strimzi.controller.cluster.operator.resource.ConfigMapOperator;
 import io.strimzi.controller.cluster.operator.resource.DeploymentConfigOperator;
