+++ cluster-controller/src/test/java/io/strimzi/controller/cluster/operator/resource/ZookeeperSetOperationsTest.java
@@ -2,13 +2,15 @@
  * Copyright 2018, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.controller.cluster.operator.assembly;
+package io.strimzi.controller.cluster.operator.resource;
 
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.EnvVar;
 import io.fabric8.kubernetes.api.model.extensions.StatefulSet;
 import io.strimzi.controller.cluster.ResourceUtils;
 import io.strimzi.controller.cluster.model.ZookeeperCluster;
+import io.strimzi.controller.cluster.operator.resource.StatefulSetDiff;
+import io.strimzi.controller.cluster.operator.resource.ZookeeperSetOperator;
 import org.junit.Before;
 import org.junit.Test;
